Fix variant value type mismatch with newer rattler-build (#68)
Fix the variant value comparison to match strings, in order to handle
argument mismatch between `used_variant` and `package_variants`. With
newer rattler-build versions, the former has actual types, while the
latter has strings, so e.g. `13 != "13"` and no variant is matched,
resulting in an `IndexError`. Add a dummy `c_compiler_version` use to
the test recipe to reproduce it.